This PR fixes alignment bug by introducing ColumnAlign field. 
Before
```
+----+------------------------------------------------------------------------------+---------------+------------+---------------+
| ID | Query_Execution_Plan                                                         | Rows_Returned | Executions | Total_Latency |
+----+------------------------------------------------------------------------------+---------------+------------+---------------+
| *0 | Distributed Union (distribution_table: Singers, split_ranges_aligned: false) | 0             | 1          | 0.05 msecs    |
| 1  | +- Local Distributed Union                                                   | 0             | 1          | 0.03 msecs    |
| 2  |    +- Serialize Result                                                       | 0             | 1          | 0.03 msecs    |
| 3  |       +- Filter Scan (seekable_key_size: 0)                                  |               |            |               |
| *4 |          +- Table Scan (Table: Singers, scan_method: Scalar)                 | 0             | 1          | 0.03 msecs    |
+----+------------------------------------------------------------------------------+---------------+------------+---------------+
```
After
```
+----+------------------------------------------------------------------------------+---------------+------------+---------------+
| ID | Query_Execution_Plan                                                         | Rows_Returned | Executions | Total_Latency |
+----+------------------------------------------------------------------------------+---------------+------------+---------------+
| *0 | Distributed Union (distribution_table: Singers, split_ranges_aligned: false) | 0             | 1          | 0.06 msecs    |
|  1 | +- Local Distributed Union                                                   | 0             | 1          | 0.04 msecs    |
|  2 |    +- Serialize Result                                                       | 0             | 1          | 0.03 msecs    |
|  3 |       +- Filter Scan (seekable_key_size: 0)                                  |               |            |               |
| *4 |          +- Table Scan (Table: Singers, scan_method: Scalar)                 | 0             | 1          | 0.03 msecs    |
+----+------------------------------------------------------------------------------+---------------+------------+---------------+
```